+++ plugins/repository-azure/qa/microsoft-azure-storage/build.gradle
@@ -26,6 +26,9 @@ import static org.elasticsearch.gradle.PropertyNormalization.IGNORE_VALUE
 
 apply plugin: 'elasticsearch.standalone-rest-test'
 apply plugin: 'elasticsearch.rest-test'
+apply plugin: 'elasticsearch.test.fixtures'
+
+testFixtures.useFixture ":test:fixtures:azure-fixture", "azure-fixture"
 
 boolean useFixture = false
 
@@ -38,18 +41,10 @@ String azureSasToken = System.getenv("azure_storage_sas_token")
 if (!azureAccount && !azureKey && !azureContainer && !azureBasePath && !azureSasToken) {
     azureAccount = 'azure_integration_test_account'
     azureKey = 'YXp1cmVfaW50ZWdyYXRpb25fdGVzdF9rZXk=' // The key is "azure_integration_test_key" encoded using base64
-    azureContainer = 'container_test'
-    azureBasePath = 'integration_test'
+    azureContainer = 'container'
+    azureBasePath = ''
     azureSasToken = ''
     useFixture = true
-}
-
-/** A task to start the fixture which emulates an Azure Storage service **/
-task azureStorageFixture(type: AntFixture) {
-    dependsOn testClasses
-    env 'CLASSPATH', "${ -> project.sourceSets.test.runtimeClasspath.asPath }"
-    executable = new File(project.runtimeJavaHome, 'bin/java')
-    args 'org.elasticsearch.repositories.azure.AzureStorageFixture', baseDir, azureContainer
 }
 
 Map<String, Object> expansions = [
@@ -77,11 +72,15 @@ testClusters.integTest {
     }
 
     if (useFixture) {
-        tasks.integTest.dependsOn azureStorageFixture
+        def azureAddress = {
+            int ephemeralPort = project(':test:fixtures:azure-fixture').postProcessFixture.ext."test.fixtures.azure-fixture.tcp.8091"
+            assert ephemeralPort > 0
+            'http://127.0.0.1:' + ephemeralPort
+        }
         // Use a closure on the string to delay evaluation until tests are executed. The endpoint_suffix is used
         // in a hacky way to change the protocol and endpoint. We must fix that.
         setting 'azure.client.integration_test.endpoint_suffix',
-                { "ignored;DefaultEndpointsProtocol=http;BlobEndpoint=http://${azureStorageFixture.addressAndPort }" }, IGNORE_VALUE
+                { "ignored;DefaultEndpointsProtocol=http;BlobEndpoint=${ -> azureAddress() }" }, IGNORE_VALUE
         String firstPartOfSeed = project.rootProject.testSeed.tokenize(':').get(0)
         setting 'thread_pool.repository_azure.max', (Math.abs(Long.parseUnsignedLong(firstPartOfSeed, 16) % 10) + 1).toString(), System.getProperty('ignore.tests.seed') == null ? DEFAULT : IGNORE_VALUE
     }
